LayoutTests/fast/css-intrinsic-dimensions/multicol-expected.txt

Test intrinsic width calculation for multicol. See http://dev.w3.org/csswg/css-sizing/#multicol-intrinsic

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS elm01.offsetWidth is 330
PASS elm02.offsetWidth is 330
PASS elm03.offsetWidth is 330
PASS elm04.offsetWidth is 630
PASS elm05.offsetWidth is 50
PASS elm06.offsetWidth is 330
PASS elm07.offsetWidth is 50
PASS elm08.offsetWidth is 630
PASS elm09.offsetWidth is 100
PASS elm10.offsetWidth is 480
PASS elm11.offsetWidth is 100
PASS elm12.offsetWidth is 630
PASS elm13.offsetWidth is 50
PASS elm14.offsetWidth is 100
PASS elm15.offsetWidth is 50
PASS elm16.offsetWidth is 200
PASS elm17.offsetWidth is 100
PASS elm18.offsetWidth is 150
PASS elm19.offsetWidth is 100
PASS elm20.offsetWidth is 200
PASS successfullyParsed is true

TEST COMPLETE
